Create R Program_Decision Tree.R
This R program generates decision trees on whether a person has a good/bad loan based on the home loan data. In step 2, I used Gini and Entropy to create the decision trees first and then used ROC curve to do the comparison.
In step 3, I used ANOVA and Poisson method to draw the decision trees and used RMSE to compare the accuracy.
In step 4, a compound method is used by multiplying the tree only considering the loan flags and the tree only considering the loan amount when it's a bad flag.